+++ changelog.md
@@ -1,3 +1,8 @@
+## 1.1.4.0
+- Fix Scriptable Migration not thread safe with many DbContexts
+- Remove SchemaDbContext type
+- EmbededScriptNameSpace parameter for ScriptableMigration constructor is mandatory
+
 ## 1.0.2.2
 - Fix auto-migration of SchemaDbContext with multiple instances